Added shell sort algorithm.

+++ insertion_sort.cpp
@@ -1,10 +1,10 @@
 #include "insertion_sort.h"
 
 void insertionSort(std::vector<HeapNode>& heap) {
-    int n = heap.size();
-    for (int i = 1; i < n; i++) {
+    long n = heap.size();
+    for (long i = 1; i < n; i++) {
         HeapNode key = heap[i];
-        int j = i - 1;
+        long j = i - 1;
 
         // Move elements that are greater than key to one position ahead
         while (j >= 0 && (heap[j].getRating() < key.getRating() ||
